wifi: iwlwifi: mvm: refactor iwl_mvm_cfg_he_sta()
This function sends the STA_HE_CTXT_CMD, which won't be used in the
new MLD API. Instead, a part of the fields of this command
will be in the new STA cmd. Put the parts that are common to both
commands in functions, which will later be used in sending the new
STA cmd.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
Link: https://lore.kernel.org/r/20230320122330.19b9b6c346ab.I0b8512eb64d8f03ff83879bafe9707f897d5b3c6@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 4df6a07551c8a70460b590e534785d5e97d1d528)