wifi: iwlwifi: mvm: refactor iwl_mvm_sta_send_to_fw()
We have a new STA cmd as part of the new MLD API. There are some
parts of sending the STA cmd which are common to both the old and
the new one. Put this parts in functions which will later be used
to send the new STA cmd.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
Link: https://lore.kernel.org/r/20230320122330.4da940bd7384.I3a66990fbffe9611b5e41f3686c2aff37ba2eb56@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 77b6a2e5f49bb3a4756a4b82f24221e62fa6ea8c)